perf: update popper.js install method
This commit updates how the auro-popover element consumed the library
popper.js code to better support ESM/SSR development environments.

NOTE: This update does not 100% solve the issues we have with popper.js.
This version of the library is deprecated and the owners of it have
moved to a new platform, Floating UI.

This element will still require upgrading to either a new library tool
or support native popover browser support.

Changes to be committed:
modified:   src/auro-popover.js
modified:   src/popover.js